FAMILY_HIST init issue with DocumentTag _get_key()
DocumentTag generates a new class for comparison primarily to manage the
type1 type2 interchangeability requirement.  However generating this new
tag ends up running into the FAMILY_HIST cleaning hack in the
FamilyHistTag __init__ function. This was not correctly designed to pass
over this hack if the indicator IS present in the lxml element. This
commit fixes that bug.